relocate weather mapper
