--- src/main/kotlin/data/weather/repository/mapper/WeatherMapper.kt
@@ -1,8 +1,0 @@
-package data.weather.repository.mapper
-
-import data.weather.model.DtoWeather
-import logic.model.Weather
-
-interface WeatherMapper {
-    fun mapDtoToWeather(dtoWeather: DtoWeather): Weather
-}
